refactor : 컴포넌트 style, contain 코드 분리
