+++ src/components/Main/Card/contain.ts
@@ -1,0 +1,3 @@
+import UserProfile from "./UserProfile";
+
+export { UserProfile };
