+++ src/components/Main/contain.ts
@@ -1,0 +1,4 @@
+import Card from './Card/Card';
+
+
+export { Card };
