+++ src/components/Navbar/MainNavBar/MainNavBar.tsx
@@ -1,15 +1,15 @@
-import styled from "@emotion/styled";
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { useRecoilState } from "recoil";
-import * as icons from "../../lib/asset/svg";
-import { ICON_MIDDLE_SIZE, ICON_SMALL_SIZE } from "../../lib/styles/size";
-import { logout } from "../../lib/utils/loginCheck";
+import * as icons from "../../../lib/asset/svg";
+import { ICON_MIDDLE_SIZE, ICON_SMALL_SIZE } from "../../../lib/styles/size";
+import { logout } from "../../../lib/utils/loginCheck";
 import {
   createWorkSpaceModalStateAtom,
   profileModalStateAtom,
-} from "../../states/main";
-import CustomIcon from "../common/CustomIcon";
+} from "../../../states/main";
+import CustomIcon from "../../common/CustomIcon/CustomIcon";
+import * as S from "./style";
 
 const MainNavBar: React.FC<any> = () => {
   const [createWorkSpaceModalState, setCreateWorkSpaceModalState] =
@@ -17,20 +17,19 @@ const MainNavBar: React.FC<any> = () => {
   const [profileModalState, setProfileModalState] = useRecoilState<boolean>(
     profileModalStateAtom
   );
-  const navigate = useNavigate();
 
   return (
     <>
-      <Middle>
+      <S.Middle>
         <CustomIcon
           icon={icons.PlusIcon}
           height={ICON_MIDDLE_SIZE}
           onClick={() => {
-            console.log(createWorkSpaceModalState);
+            // console.log(createWorkSpaceModalState);
             setCreateWorkSpaceModalState(!createWorkSpaceModalState);
           }}
         />
-      </Middle>
+      </S.Middle>
       <CustomIcon
         icon={icons.ProfileIconSvg}
         height={ICON_SMALL_SIZE}
@@ -53,13 +52,3 @@ const MainNavBar: React.FC<any> = () => {
   );
 };
 export default MainNavBar;
-
-const Middle = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  grid-column: 3;
-  & > div {
-    margin-left: 10px;
-  }
-`;
