+++ src/components/Navbar/MainNavBar/style.ts
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+
+export const Middle = styled.div`
+display: flex;
+width: 100%;
+justify-content: center;
+grid-column: 3;
+& > div {
+  margin-left: 10px;
+}
+`;
