+++ src/components/Navbar/contain.ts
@@ -1,0 +1,6 @@
+import MainNavBar from "./MainNavBar/MainNavBar";
+import WorkSpaceNavBar from "./WorkSpaceNavbar/WorkSpaceNavBar"
+
+
+
+export { MainNavBar, WorkSpaceNavBar };
