+++ src/components/Navbar/style.ts
@@ -1,0 +1,17 @@
+import styled from "@emotion/styled";
+
+export const NavBarLayout = styled.div`
+z-index: 3;
+position: relative;
+top: 0;
+left: 0;
+width: 100%;
+height: 50px;
+background-color: #ffffff;
+display: grid;
+padding: 0 10px;
+box-sizing: border-box;
+align-items: center;
+grid-template-columns: 40px 40px 1fr 40px 40px;
+overflow: hidden;
+`;
