--- src/components/common/Loading/style.css
@@ -1,82 +1,0 @@
-
-.loader {
-	position: relative;
-	display: inline-block;
-	width: 50px;
-	height: 50px;
-	vertical-align: middle;
-}
-
-
-.loader-double {
-	border-radius: 50px;
-	border: 6px solid transparent;
-	border-top-color: #fff;
-	border-bottom-color: #fff;
-	-webkit-animation: spin 1.5s linear infinite;
-	animation: spin 1.5s linear infinite;
-}
-
-.loader-double:before,
-.loader-double:after {
-	content: '';
-	position: absolute;
-	top: 5px; 
-	left: 5px;
-	bottom: 5px;
-	right: 5px;
-	border-radius: 50px;
-	border: 6px solid transparent;
-	border-top-color: #fff;
-	border-bottom-color: #fff;
-	filter: alpha(opacity=6);
-	     -khtml-opacity: .6;
-	       -moz-opacity: .6;
-	            opacity: .6;
-	-webkit-animation: spinreverse 2s linear infinite;
-	animation: spinreverse 2s linear infinite;
-}
-
-.loader-double:before {
-	top: 15px; 
-	left: 15px;
-	bottom: 15px;
-	right: 15px;
-	-webkit-animation: spinreverse 3s linear infinite;
-	animation: spinreverse 3s linear infinite;
-}
-
-/*	$Loader Multiple circle
-	========================================================================== */
-
-@-webkit-keyframes grow {
-	0% { -webkit-transform: scaleY(0); transform: scaleY(0); opacity: 0;}
-	50% { -webkit-transform: scaleY(1); transform: scaleY(1); opacity: 1;}
-	100% { -webkit-transform: scaleY(0); transform: scaleY(0); opacity: 0;}
-}
-
-@keyframes grow {
-	0% { -webkit-transform: scaleY(0); transform: scaleY(0); opacity: 0;}
-	50% { -webkit-transform: scaleY(1); transform: scaleY(1); opacity: 1;}
-	100% { -webkit-transform: scaleY(0); transform: scaleY(0); opacity: 0;}
-}
-
-@-webkit-keyframes spin {
-  	0%{ -webkit-transform: rotate(0deg); tranform: rotate(0deg);}
-	100%{ -webkit-transform: rotate(360deg); tranform: rotate(360deg);}
-}
-
-@keyframes spin {
-	0%{ -webkit-transform: rotate(0deg); transform: rotate(0deg);}
-	100%{ -webkit-transform: rotate(360deg); transform: rotate(360deg);}
-}
-
-@-webkit-keyframes spinreverse {
-  	0%{ -webkit-transform: rotate(0deg); tranform: rotate(0deg);}
-	100%{ -webkit-transform: rotate(-360deg); tranform: rotate(-360deg);}
-}
-
-@keyframes spinreverse {
-	0%{ -webkit-transform: rotate(0deg); transform: rotate(0deg);}
-	100%{ -webkit-transform: rotate(-360deg); transform: rotate(-360deg);}
-}
