+++ src/components/common/Modal/CreateWorkSpace/CreateWorkSpaceModal.tsx
@@ -1,4 +1,3 @@
-import styled from "@emotion/styled";
 import React, {
   ChangeEvent,
   ChangeEventHandler,
@@ -6,12 +5,16 @@ import React, {
   useState,
 } from "react";
 import { useRecoilState, useSetRecoilState } from "recoil";
-import { getUserInfo } from "../../../lib/api/user";
-import { createWorkSpace } from "../../../lib/api/workSpace";
-import { Button } from "../../../lib/styles/button";
-import { User } from "../../../lib/types/user";
-import { createWorkSpaceModalStateAtom, userAtom } from "../../../states/main";
-import Modal from "./Modal";
+import { getUserInfo } from "../../../../lib/api/user";
+import { createWorkSpace } from "../../../../lib/api/workSpace";
+import { Button } from "../../../../lib/styles/button";
+import { User } from "../../../../lib/types/user";
+import {
+  createWorkSpaceModalStateAtom,
+  userAtom,
+} from "../../../../states/main";
+import Modal from "../Modal";
+import * as S from "./style";
 export interface CreateWorkSpaceModalProps {}
 
 const CreateWorkSpaceModal = ({}: CreateWorkSpaceModalProps) => {
@@ -49,58 +52,20 @@ const CreateWorkSpaceModal = ({}: CreateWorkSpaceModalProps) => {
 
   return (
     <Modal state={createWorkSpaceModalState} onHandleState={onHandleState}>
-      <Wrapper>
+      <S.Wrapper>
         <div>로고</div>
-        <Title>
-          <Text>워크스페이스 이름</Text>
-          <TitleInput
+        <S.Title>
+          <S.Text>워크스페이스 이름</S.Text>
+          <S.TitleInput
             value={value}
             onChange={onChangeValue}
             onKeyPress={handleEnter}
           />
-        </Title>
+        </S.Title>
         <Button onClick={handleCreateWorkSpace}>워크스페이스 생성</Button>
-      </Wrapper>
+      </S.Wrapper>
     </Modal>
   );
 };
 
 export default CreateWorkSpaceModal;
-
-const Wrapper = styled.div`
-  width: 50vw;
-  height: 50vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const Title = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  flex: 1;
-`;
-
-const Text = styled.div`
-  font-size: 1.3rem;
-  margin-bottom: 1rem;
-`;
-
-const TitleInput = styled.input`
-  font-size: 1.3rem;
-  border: 0;
-  width: 80%;
-  border-bottom: 1px solid black;
-  text-align: center;
-  &:active {
-    /* background-color: black; */
-  }
-  &:focus {
-    background-color: #eaeaea80;
-    border-radius: 5px 5px 0px 0px;
-    outline: 0px;
-    transition: ease 0.3s;
-  }
-`;
