+++ src/components/common/Modal/InviteModal/contain.ts
@@ -1,0 +1,4 @@
+import Member from './Member/Member';
+
+
+export { Member };
