+++ src/components/common/Modal/Profile/contain.ts
@@ -1,0 +1,3 @@
+import EditContent from "./EditContent/EditContent";
+
+export { EditContent };
