+++ src/lib/types/resizeableTab.ts
@@ -1,0 +1,10 @@
+export type tabPosition = "left" | "right" | "bottom";
+
+
+export type tabState = {
+    onHandling: boolean;
+    position: tabPosition;
+    width: string;
+    height: string;
+  };
+
